Moved WaitNotifyService to impl
The WaitNotifyService is an implementation detail and is not used
by the SPI user. So therefor it is moved. As part of the moving
the InternalWaitNotifyService has been merged into the WaitNotifyService.

No Enterprise change needed.
