+++ hazelcast/src/main/java/com/hazelcast/spi/impl/waitnotifyservice/package-info.java
@@ -15,6 +15,6 @@
  */
 
 /**
- * Contains the internal API code for the {@link com.hazelcast.spi.WaitNotifyService}.
+ * Contains the internal API code for the {@link com.hazelcast.spi.impl.waitnotifyservice.WaitNotifyService}.
  */
 package com.hazelcast.spi.impl.waitnotifyservice;
